Moved PointInstanceType into points/types
