+++ packages/replay-next/components/console/LoggablesContext.tsx
@@ -11,7 +11,7 @@ import {
 import { ConsoleFiltersContext } from "replay-next/src/contexts/ConsoleFiltersContext";
 import { FocusContext } from "replay-next/src/contexts/FocusContext";
 import { Context as ConsolePointsContext } from "replay-next/src/contexts/points/ConsolePointsContext";
-import { PointInstance } from "replay-next/src/contexts/points/PointsContext";
+import { PointInstance } from "replay-next/src/contexts/points/types";
 import { TerminalContext, TerminalExpression } from "replay-next/src/contexts/TerminalContext";
 import { EventLog, getEventTypeEntryPointsSuspense } from "replay-next/src/suspense/EventsCache";
 import { UncaughtException, getExceptionsSuspense } from "replay-next/src/suspense/ExceptionsCache";
