+++ packages/replay-next/src/utils/loggables.ts
@@ -1,7 +1,7 @@
 import { ExecutionPoint } from "@replayio/protocol";
 
 import { Loggable } from "replay-next/components/console/LoggablesContext";
-import { PointInstance } from "replay-next/src/contexts/points/PointsContext";
+import { PointInstance } from "replay-next/src/contexts/points/types";
 import { TerminalExpression } from "replay-next/src/contexts/TerminalContext";
 import { EventLog } from "replay-next/src/suspense/EventsCache";
 import { UncaughtException } from "replay-next/src/suspense/ExceptionsCache";
